PokeDex_App/requirements.txt
pypokedex==1.6.0
Pillow==8.2.0